+++ src/dipdup/datasources/metadata/enums.py
@@ -2,6 +2,6 @@ from enum import Enum
 
 
 class MetadataNetwork(Enum):
-    mainnet = 'mainnet'
-    # FIXME: ...2net?
     hangzhounet = 'hangzhounet'
+    ithacanet = 'ithacanet'
+    mainnet = 'mainnet'